Renamed the existing model file to indicate development.
Removed the saved ruleset from the development model file.
Completed the following in order to have the model solve correctly when switching back to 24 MS mode from MTOM mode.

Added a BHOPSParkerOutflow_Initialization Expression Slot.
Added an action in the switch to 24 MS mode script which evaluates the new expression slot above.
Disabled setting the BHOPSParkerOutflow to input in the change to 24 MS script
Added initialization rule to solve for the HavasuOutflow.BHOPSParkerOutflow slot in 24 MS mode if NaN. This uses the expression slot calculated when switching modes.