#33
FG Spring and Baseflow A-J Volume() modified in CRMMS_V1.5.9000.rls: Removed March from the Jan-Feb period of simulated uncertainty when the April-July volume forecast volume is calculated as the average volume for that period multiplied by the ratio of the current year's forecast volume from the current month through April to the average for the same period. This is done to incorporate some uncertainty and ensure operations were sufficiently conservative should the forecast change. This created an issue when we CRMMS rules did the first pass (WHAM) to solve the Max probable 24MS. A conservative estimation of releases in March can make FG not release enough in April to make the May 1 target. In the future we might revisit how both model simulate uncertainty in the FG forecast, if CRSS should be brought more into line with what is done in CRSS at Lake Powell. Also if CRMMS should simulate this uncertainty at all