February 2022 CRMMS ESP Run. This includes the model file, ruleset, and output from the February 2022 ESP run.
An update was made to the ruleset during the February run. A small logic change was applied to MWD diversion adjustment rule to make sure the final diversion set in year 1 of the ESP run was more representative of the operational schedule distribution. The ruleset name was incremented from CRMMS_V1.10.rls to CRMMS_V1.10.1.rls.