Added slots to track DROA operations
This commit adds slots to track DROA operations. None of the series have any data and scalars have provisional data.

*************
Important Note:
Accounting is now enabled on the accounting menu.  DO NOT DISABLE IT.  It should not cause any issues as long as the controller stays on Rulebased Simulation and is not changed to an accounting version.
*************|

For DROA slots added:
- Slots use camel case in all names. DROA is separated from other capital letters with an underscore.
- Each new slot is in a DROA slot group.
- Slot descriptions indicate what each value means.

Initially, I said many of these would be expression slots but I've changed them to series to be set by a rules for two reasons: 1. They can solve when needed within a timestep (Expr slots only solve at the start or end of a timestep or run), which will be important in the future in Ensemble or MTOM mode.   2. There would be a lot of new things required like an expr slot function set or global function set. For incremental development, this approach is better and more self contained.

New slots on ICUs, ResData are:
-- DROA_Flag - scalar.  This slot triggers whether DROA is in effect for this reservoir.
-- PoolElevationWithDROA_Ops - series (to be set by rule --> same as Res.Pool Elevation)
-- PoolElevationWithoutDROA_Ops - series (to be set by rule)
-- ReleaseWithDROA_Ops - series (input)
-- ReleaseWithoutDROA_Ops - series (input)
-- ReleaseChangeDueToDROA_Ops - series (to be set by rule --> difference of above 2)
-- StorageWithDROA_Ops - series (to be set by rule --> same as Res.Storage)
-- StorageWithoutDROA_Ops - series (to be set by rule --> res.storage + Powell^ResDROA.Storage)

New Slots on PowellData object are:
-- DROA_FutilityThresholdElevation - In 24Mo study mode, for Plotting only.
-- DROA_RecoveryThresholdElevation - In 24Mo study mode, for Plotting only.
-- DROA_StorageTotal - series (to be set by rule) summing up all DROA storage accts
-- DROA_TargetElevation - In 24Mo study mode, for Plotting only.
-- PoolElevationWithDROA_Ops - series (to be set by rule --> same as Res.Pool Elevation)
-- PoolElevationWithoutDROA_Ops - series (to be set by rule)
-- StorageWithoutDROA_Ops - series (to be set by rule --> res.storage + Powell^ResDROA.Storage)
-- StorageWithDROA_Ops - series (to be set by rule --> same as Res.Storage)