CRMMS DROA project - new SCT for DROA Ops
This adds in SCT for use in DROA operations.  It has one tab per UIU and one for Powell.
It shows Inflows, outflows, storage, pool elevation and then some of those same variables but "without DROA".
Also, it shows Account storage for DROA water in Powell.
The scalar slots tab has the three UIU "DROA Ops" flags that are used to indicate if DROA is to be considered for a particular res.