CRMMS DROA - changes to support automatic recovery, if desired. Minor SCT changes.
This commit adds in a rule to automatically recover DROA as the UIU gets full.
For each reservoir, there is an elevation or storage level that triggers recovery.  As the sum of total storage and Powell^ResDROA storage gets above that level, then that represents recovery. That volume is transferred from Powell^ResDROA to Powell^System in the "DROA Recovery Transfers Due to Elevation" rule does this.  I also updated many of the guide curves with dummy or test data, often from PWRE's DROA model. These are provisional.  As part of the demo, Reclamation said this approach is still controversial. If not desired, the rule can be turned off or deleted.  Guide curve slots can also be removed from UIUResData.

This commit also moves all DROA rules to execute in run cycle 2 as that is when releases from UIU are made in ESP mode. So looking forward, this is where the DROA rules should go.

Made minor changes to the DROA Ops SCT after the demo on Friday.