DROA Modeling task - Changed a few slots names to match the convention and other general cleanup
This commit changes a few slot names to match the convention: camel case but underscores when a capital letter is next to DROA.
I also updated the SCT with the new slot references.

I cleaned up the comments on the rules and functions to remove any reference to PWRE comments.   I changed a few minor rule items to reorganize assignments and use NOT isInput on initial PE.